Started profiling test suite
There appears to be a difference between running the test suite from
nosetests and running it interactively. The answers are slightly
difference (hence the additional limits to only match unit tests to 6
decimal places) and the tests run *much* faster when run interactively.
It's almost like I'm importing the python version from nose, and the
cython version when I run from the command line.